requirements.txt
yfinance
pandas
streamlit
requests
python-dotenv
